Refactor Moondream3 model components for improved type handling and performance
- Updated language.py to ensure consistent data types in computations, reducing unnecessary casts.
- Introduced _gather_sort and _scatter_unsort functions for efficient token sorting and restoration in MoEMLP.
- Enhanced moondream3.py to maintain input data type consistency for pixel values.
- Modified vision.py to allocate pooled tensors with the correct data type, improving memory management.

+++ mlx_vlm/models/moondream3/vision.py
@@ -181,7 +181,7 @@ class VisionModel(nn.Module):
 
         # Simple approach: use strided slicing + averaging via reshape when possible
         # For arbitrary sizes, use a loop-based approach
-        pooled = mx.zeros((grid_size, grid_size, D))
+        pooled = mx.zeros((grid_size, grid_size, D), dtype=full_grid.dtype)
         for i in range(grid_size):
             h_start = int(round(i * pool_h))
             h_end = int(round((i + 1) * pool_h))
